+++ src/task.c
@@ -568,7 +568,7 @@ void ABTI_task_print(ABTI_task *p_task, FILE *p_os, int indent)
     }
 
     ABTI_xstream *p_xstream = p_task->p_xstream;
-    size_t xstream_rank = p_xstream ? p_xstream->rank : 0;
+    uint64_t xstream_rank = p_xstream ? p_xstream->rank : 0;
     char *state;
     switch (p_task->state) {
         case ABT_TASK_STATE_CREATED:    state = "CREATED"; break;
